+++ pkg/util/validation/validate_test.go
@@ -183,20 +183,20 @@ func TestValidateLabels(t *testing.T) {
 		assert.Equal(t, c.err, err, "wrong error")
 	}
 
-	validateMetrics.DiscardedSamples.WithLabelValues("random reason", "different user").Inc()
+	validateMetrics.DiscardedSamples.WithLabelValues("random reason", "different user", SampleTypeFloat).Inc()
 
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="label_invalid",user="testUser"} 1
-			cortex_discarded_samples_total{reason="label_name_too_long",user="testUser"} 1
-			cortex_discarded_samples_total{reason="label_value_too_long",user="testUser"} 1
-			cortex_discarded_samples_total{reason="max_label_names_per_series",user="testUser"} 1
-			cortex_discarded_samples_total{reason="metric_name_invalid",user="testUser"} 1
-			cortex_discarded_samples_total{reason="missing_metric_name",user="testUser"} 1
-			cortex_discarded_samples_total{reason="labels_size_bytes_exceeded",user="testUser"} 1
+			cortex_discarded_samples_total{reason="label_invalid",type="float",user="testUser"} 1
+			cortex_discarded_samples_total{reason="label_name_too_long",type="float",user="testUser"} 1
+			cortex_discarded_samples_total{reason="label_value_too_long",type="float",user="testUser"} 1
+			cortex_discarded_samples_total{reason="max_label_names_per_series",type="float",user="testUser"} 1
+			cortex_discarded_samples_total{reason="metric_name_invalid",type="float",user="testUser"} 1
+			cortex_discarded_samples_total{reason="missing_metric_name",type="float",user="testUser"} 1
+			cortex_discarded_samples_total{reason="labels_size_bytes_exceeded",type="float",user="testUser"} 1
 
-			cortex_discarded_samples_total{reason="random reason",user="different user"} 1
+			cortex_discarded_samples_total{reason="random reason",type="float",user="different user"} 1
 	`), "cortex_discarded_samples_total"))
 
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
@@ -212,7 +212,7 @@ func TestValidateLabels(t *testing.T) {
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="random reason",user="different user"} 1
+			cortex_discarded_samples_total{reason="random reason",type="float",user="different user"} 1
 	`), "cortex_discarded_samples_total"))
 }
 
@@ -353,7 +353,7 @@ func TestValidateLabelOrder(t *testing.T) {
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="labels_not_sorted",user="testUser"} 1
+			cortex_discarded_samples_total{reason="labels_not_sorted",type="float",user="testUser"} 1
 	`), "cortex_discarded_samples_total"))
 }
 
@@ -391,7 +391,7 @@ func TestValidateLabelDuplication(t *testing.T) {
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="duplicate_label_names",user="testUser"} 2
+			cortex_discarded_samples_total{reason="duplicate_label_names",type="float",user="testUser"} 2
 	`), "cortex_discarded_samples_total"))
 }
 
@@ -670,7 +670,7 @@ func TestValidateNativeHistogram(t *testing.T) {
 			actualHistogram, actualErr := ValidateNativeHistogram(validateMetrics, limits, userID, lbls, tc.histogram)
 			if tc.expectedErr != nil {
 				require.Equal(t, tc.expectedErr.Error(), actualErr.Error())
-				require.Equal(t, float64(1), testutil.ToFloat64(validateMetrics.DiscardedSamples.WithLabelValues(tc.discardedSampleLabelValue, userID)))
+				require.Equal(t, float64(1), testutil.ToFloat64(validateMetrics.DiscardedSamples.WithLabelValues(tc.discardedSampleLabelValue, userID, SampleTypeHistogram)))
 				// Should never increment if error was returned
 				require.Equal(t, float64(0), testutil.ToFloat64(validateMetrics.HistogramSamplesReducedResolution.WithLabelValues(userID)))
 
@@ -710,7 +710,7 @@ func TestValidateMetrics_UpdateSamplesDiscardedForSeries(t *testing.T) {
 			cortex_discarded_samples_per_labelset_total{labelset="{foo=\"bar\"}",reason="dummy",user="user"} 100
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 
 	v.updateSamplesDiscardedForSeries(userID, "out-of-order", limits, labels.FromMap(map[string]string{"foo": "baz"}), 1)
@@ -721,8 +721,8 @@ func TestValidateMetrics_UpdateSamplesDiscardedForSeries(t *testing.T) {
 			cortex_discarded_samples_per_labelset_total{labelset="{foo=\"baz\"}",reason="out-of-order",user="user"} 1
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
-			cortex_discarded_samples_total{reason="out-of-order",user="user"} 1
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
+			cortex_discarded_samples_total{reason="out-of-order",type="float",user="user"} 1
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 
 	// Match default partition.
@@ -735,9 +735,9 @@ func TestValidateMetrics_UpdateSamplesDiscardedForSeries(t *testing.T) {
 			cortex_discarded_samples_per_labelset_total{labelset="{}",reason="too-old",user="user"} 1
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
-			cortex_discarded_samples_total{reason="out-of-order",user="user"} 1
-			cortex_discarded_samples_total{reason="too-old",user="user"} 1
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
+			cortex_discarded_samples_total{reason="out-of-order",type="float",user="user"} 1
+			cortex_discarded_samples_total{reason="too-old",type="float",user="user"} 1
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 }
 
@@ -761,7 +761,7 @@ func TestValidateMetrics_UpdateLabelSet(t *testing.T) {
 		},
 	}
 
-	v.updateSamplesDiscarded(userID, "dummy", limits, 100)
+	v.updateSamplesDiscarded(userID, "dummy", SampleTypeFloat, limits, 100)
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 			# HELP cortex_discarded_samples_per_labelset_total The total number of samples that were discarded for each labelset.
 			# TYPE cortex_discarded_samples_per_labelset_total counter
@@ -770,7 +770,7 @@ func TestValidateMetrics_UpdateLabelSet(t *testing.T) {
 			cortex_discarded_samples_per_labelset_total{labelset="{}",reason="dummy",user="user"} 100
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 
 	// Remove default partition.
@@ -786,7 +786,7 @@ func TestValidateMetrics_UpdateLabelSet(t *testing.T) {
 			cortex_discarded_samples_per_labelset_total{labelset="{foo=\"baz\"}",reason="dummy",user="user"} 100
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 
 	// Remove limit 1.
@@ -801,7 +801,7 @@ func TestValidateMetrics_UpdateLabelSet(t *testing.T) {
 			cortex_discarded_samples_per_labelset_total{labelset="{foo=\"bar\"}",reason="dummy",user="user"} 100
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 
 	// Remove user.
@@ -810,6 +810,6 @@ func TestValidateMetrics_UpdateLabelSet(t *testing.T) {
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 			# HELP cortex_discarded_samples_total The total number of samples that were discarded.
 			# TYPE cortex_discarded_samples_total counter
-			cortex_discarded_samples_total{reason="dummy",user="user"} 100
+			cortex_discarded_samples_total{reason="dummy",type="float",user="user"} 100
 	`), "cortex_discarded_samples_total", "cortex_discarded_samples_per_labelset_total"))
 }